+++ R/find_response.R
@@ -55,6 +55,16 @@ find_response.model_fit <- function(x, combine = TRUE, ...) {
 
 
 #' @export
+find_response.bfsl <- function(x, combine = TRUE, ...) {
+  resp <- find_response(x, combine = combine)
+  if (is.null(resp)) {
+    resp <- "y"
+  }
+  resp
+}
+
+
+#' @export
 find_response.selection <- function(x, combine = TRUE, ...) {
   f <- find_formula(x, verbose = FALSE)
   resp <- c(