+++ R/get_parameters.R
@@ -125,6 +125,21 @@ get_parameters.model_fit <- function(x, ...) {
 
 
 #' @export
+get_parameters.bsfl <- function(x, ...) {
+  cf <- stats::coef(x)
+
+  params <- data.frame(
+    Parameter = rownames(cf),
+    Estimate = unname(cf[, "Estimate"]),
+    stringsAsFactors = FALSE,
+    row.names = NULL
+  )
+
+  .remove_backticks_from_parameter_names(params)
+}
+
+
+#' @export
 get_parameters.selection <- function(x, component = c("all", "selection", "outcome", "auxiliary"), ...) {
   component <- match.arg(component)
   s <- summary(x)